CIV-10455 Apply changes from reviews
