+++ src/main/common/models/claim.ts
@@ -82,7 +82,6 @@ import {MediationUploadDocumentsCCD} from 'models/mediation/uploadDocuments/uplo
 import {CCDHelpWithFeesDetails} from 'models/ccdResponse/ccdHelpWithFeesDetails';
 import {DirectionQuestionnaireType} from 'models/directionsQuestionnaire/directionQuestionnaireType';
 import {GeneralApplication} from './generalApplication/GeneralApplication';
-import {CuiJudgmentPaidInFull} from 'models/judgmentOnline/cuiJudgmentPaidInFull';
 
 export class Claim {
   resolvingDispute: boolean;
@@ -171,7 +170,6 @@ export class Claim {
   orderDocumentId?: string;
   claimantEvidence: ClaimantEvidence;
   defendantResponseDocuments: SystemGeneratedCaseDocuments[];
-  joJudgmentPaidInFull?: CuiJudgmentPaidInFull;
   // Index signature to allow dynamic property access
   [key: string]: any;
