+++ src/main/common/models/judgmentOnline/joClaim.ts
@@ -1,0 +1,6 @@
+import {CuiJudgmentPaidInFull} from 'models/judgmentOnline/cuiJudgmentPaidInFull';
+import {Claim} from 'models/claim';
+
+export class JoClaim extends Claim {
+  joJudgmentPaidInFull?: CuiJudgmentPaidInFull;
+}
