+++ src/main/services/translation/judgmentOnline/convertToCCDjudgmentPaidInFull.ts
@@ -1,5 +1,6 @@
 import {Claim} from 'models/claim';
 import {CCDJudgmentPaidInFull, JoJudgmentPaidInFull} from 'models/judgmentOnline/ccdJudgmentPaidInFull';
+import {JoClaim} from 'models/judgmentOnline/joClaim';
 
 export const toCCDjudgmentPaidInFull = (claim: Claim) : CCDJudgmentPaidInFull => {
   return {
@@ -7,9 +8,9 @@ export const toCCDjudgmentPaidInFull = (claim: Claim) : CCDJudgmentPaidInFull =>
   };
 };
 
-export const toCCDjudgmentPaidInFullDetails = (claim: Claim) : JoJudgmentPaidInFull => {
+export const toCCDjudgmentPaidInFullDetails = (joClaim: JoClaim) : JoJudgmentPaidInFull => {
   return {
-    dateOfFullPaymentMade: claim.joJudgmentPaidInFull ? claim.joJudgmentPaidInFull.dateOfFullPaymentMade : null,
-    confirmFullPaymentMade: claim.joJudgmentPaidInFull?.confirmFullPaymentMade ? ['CONFIRMED'] : null,
+    dateOfFullPaymentMade: joClaim.joJudgmentPaidInFull ? joClaim.joJudgmentPaidInFull.dateOfFullPaymentMade : null,
+    confirmFullPaymentMade: joClaim.joJudgmentPaidInFull?.confirmFullPaymentMade ? ['CONFIRMED'] : null,
   };
 };
