+++ src/main/views/features/judgmentOnline/judgment-online-confirmation-page.njk
@@ -1,4 +1,4 @@
-{% extends "claim-details-tpl.njk" %}
+{% extends "claim-details-tpl-dashboard.njk" %}
 {% from "govuk/components/panel/macro.njk" import govukPanel %}
 {% from "../../macro/page-title.njk" import setPageTitle %}
 {% from "../dashboard/item-content.njk" import itemContent %}
